requirements.txt
pylint
requests
websocket-client==0.56
pandas
numpy
colorama
mysql.connector